refactor: restructure path
